# specs/001-filtering-system/contracts/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Tech Hub Filtering API",
        "version": "1.0.0",
        "description": "API endpoints for content filtering with tags and date ranges"
    },
    "servers": [
        {
            "url": "https://localhost:5001",
            "description": "Local development (API)"
        },
        {
            "url": "https://tech.hub.ms/api",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/api/tags/cloud": {
            "get": {
                "summary": "Get tag cloud for specified scope",
                "description": "Returns top tags with usage counts and size categories for tag cloud rendering",
                "operationId": "getTagCloud",
                "tags": [
                    "Tags"
                ],
                "parameters": [
                    {
                        "name": "scope",
                        "in": "query",
                        "description": "Scope of tag cloud calculation",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "homepage",
                                "section",
                                "collection",
                                "content"
                            ]
                        },
                        "example": "section"
                    },
                    {
                        "name": "section",
                        "in": "query",
                        "description": "Section name (required for section/collection/content scopes)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "ai"
                    },
                    {
                        "name": "collection",
                        "in": "query",
                        "description": "Collection name (required for collection scope)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "news"
                    },
                    {
                        "name": "slug",
                        "in": "query",
                        "description": "Content item slug (required for content scope)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2024-06-07-GitHub-Copilot-Levels-of-enlightenment"
                    },
                    {
                        "name": "maxTags",
                        "in": "query",
                        "description": "Maximum number of tags to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 100,
                            "default": 20
                        }
                    },
                    {
                        "name": "minUses",
                        "in": "query",
                        "description": "Minimum usage count for tags",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 1000,
                            "default": 5
                        }
                    },
                    {
                        "name": "lastDays",
                        "in": "query",
                        "description": "Include only content from last N days (null = all time)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 365,
                            "nullable": true,
                            "default": 90
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully calculated tag cloud",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TagCloudItem"
                                    }
                                },
                                "example": [
                                    {
                                        "tag": "AI",
                                        "count": 42,
                                        "size": "Large"
                                    },
                                    {
                                        "tag": "Azure",
                                        "count": 35,
                                        "size": "Large"
                                    },
                                    {
                                        "tag": "GitHub Copilot",
                                        "count": 28,
                                        "size": "Medium"
                                    },
                                    {
                                        "tag": "DevOps",
                                        "count": 15,
                                        "size": "Medium"
                                    },
                                    {
                                        "tag": ".NET",
                                        "count": 8,
                                        "size": "Small"
                                    }
                                ]
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "error": "Section name required for 'section' scope"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tags/all": {
            "get": {
                "summary": "Get all available tags with counts",
                "description": "Returns all tags in the specified scope, sorted alphabetically, for dropdown filter",
                "operationId": "getAllTags",
                "tags": [
                    "Tags"
                ],
                "parameters": [
                    {
                        "name": "section",
                        "in": "query",
                        "description": "Section scope (optional)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "ai"
                    },
                    {
                        "name": "collection",
                        "in": "query",
                        "description": "Collection scope (optional, requires section)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "news"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved all tags",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AllTagsResponse"
                                },
                                "example": {
                                    "tags": [
                                        {
                                            "name": ".NET",
                                            "count": 8
                                        },
                                        {
                                            "name": "AI",
                                            "count": 42
                                        },
                                        {
                                            "name": "Azure",
                                            "count": 35
                                        },
                                        {
                                            "name": "DevOps",
                                            "count": 15
                                        },
                                        {
                                            "name": "GitHub Copilot",
                                            "count": 28
                                        }
                                    ],
                                    "totalTagCount": 5
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "TagCloudItem": {
                "type": "object",
                "required": [
                    "tag",
                    "count",
                    "size"
                ],
                "properties": {
                    "tag": {
                        "type": "string",
                        "description": "Tag name",
                        "example": "AI"
                    },
                    "count": {
                        "type": "integer",
                        "description": "Usage count",
                        "example": 42
                    },
                    "size": {
                        "type": "string",
                        "enum": [
                            "Small",
                            "Medium",
                            "Large"
                        ],
                        "description": "Visual size category (quantile-based)",
                        "example": "Large"
                    }
                }
            },
            "AllTagsResponse": {
                "type": "object",
                "required": [
                    "tags",
                    "totalTagCount"
                ],
                "properties": {
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TagWithCount"
                        },
                        "description": "All tags sorted alphabetically"
                    },
                    "totalTagCount": {
                        "type": "integer",
                        "description": "Total number of unique tags",
                        "example": 127
                    }
                }
            },
            "TagWithCount": {
                "type": "object",
                "required": [
                    "name",
                    "count"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Tag name",
                        "example": "AI"
                    },
                    "count": {
                        "type": "integer",
                        "description": "Number of items with this tag",
                        "example": 42
                    }
                }
            },
            "ErrorResponse": {
                "type": "object",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "string",
                        "description": "Error message",
                        "example": "Invalid filter parameters"
                    },
                    "details": {
                        "type": "object",
                        "nullable": true,
                        "description": "Additional error details",
                        "additionalProperties": true
                    }
                }
            }
        }
    }
}